requirements.txt
nomkl
sbi==0.21.0
neo
scikit-learn==1.1.3
joblib==1.3.2
elephant
mat73
func_timeout==4.3.5
pandas==1.4.2
scipy==1.10.0
matplotlib==3.7.3
tqdm
pyyaml
h5py==3.6.0